Table of Genes in a long format

Transformation of a gene expression matrix into a 3 column table.

Usage

convert_to_long(the_dataset)

Arguments

the_dataset

This is a dataset often in a wide format.

Value

gene_data

Author

Oriade Latifah Simpson

Examples

data("data_two")
my_data_two
#> # A tibble: 49 × 7,130
#>    value  gene_1 gene_2 gene_3 gene_4 gene_5 gene_6 gene_7 gene_8  gene_9
#>    <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl>
#>  1     1 -0.204  -0.788  0.657  1.69  -0.867 -0.419 -0.867 -0.298  1.58  
#>  2     1 -0.155  -0.830 -0.159 -0.830 -0.830 -0.830 -0.830 -0.830 -0.526 
#>  3     1  0.539  -0.849  0.661  1.16  -0.849 -0.849 -0.849 -0.628  0.0564
#>  4     1 -0.251  -0.820  1.50   1.95  -0.820 -0.820 -0.820 -0.796  1.23  
#>  5     1  0.0784 -0.828  0.839  1.61  -0.553 -0.116 -0.828 -0.828  0.934 
#>  6     1 -0.170  -0.788  1.44   0.836 -0.788 -0.788 -0.788 -0.788 -0.170 
#>  7     0  0.478   0.727  0.462  1.11  -0.318  0.583 -0.888  0.849  0.811 
#>  8     0  1.27    0.647  1.14   1.40  -0.746 -0.944 -0.944  0.555  0.960 
#>  9     0 -0.0352 -0.851 -0.851  0.128 -0.851 -0.851 -0.851 -0.851 -0.159 
#> 10     0  0.246  -0.855  0.118  1.01  -0.855 -0.855 -0.855 -0.855 -0.282 
#> # … with 39 more rows, and 7,120 more variables: gene_10 <dbl>, gene_11 <dbl>,
#> #   gene_12 <dbl>, gene_13 <dbl>, gene_14 <dbl>, gene_15 <dbl>, gene_16 <dbl>,
#> #   gene_17 <dbl>, gene_18 <dbl>, gene_19 <dbl>, gene_20 <dbl>, gene_21 <dbl>,
#> #   gene_22 <dbl>, gene_23 <dbl>, gene_24 <dbl>, gene_25 <dbl>, gene_26 <dbl>,
#> #   gene_27 <dbl>, gene_28 <dbl>, gene_29 <dbl>, gene_30 <dbl>, gene_31 <dbl>,
#> #   gene_32 <dbl>, gene_33 <dbl>, gene_34 <dbl>, gene_35 <dbl>, gene_36 <dbl>,
#> #   gene_37 <dbl>, gene_38 <dbl>, gene_39 <dbl>, gene_40 <dbl>, …
my_long_data <- convert_to_long(my_data_two)
my_long_data
#> # A tibble: 349,321 × 3
#>    value Gene    expression_level
#>    <dbl> <chr>              <dbl>
#>  1     1 gene_1            -0.204
#>  2     1 gene_2            -0.788
#>  3     1 gene_3             0.657
#>  4     1 gene_4             1.69 
#>  5     1 gene_5            -0.867
#>  6     1 gene_6            -0.419
#>  7     1 gene_7            -0.867
#>  8     1 gene_8            -0.298
#>  9     1 gene_9             1.58 
#> 10     1 gene_10            1.48 
#> # … with 349,311 more rows